refactor(playback): extract useDrawerPlayback hook, redirect helper, clean dead memos (#2232)
Addresses code review findings:
- extracted useDrawerPlayback alongside use-wall-confirm-fallback / use-drawer-url-sync
- peer-sync subscription is now unconditional so opening the drawer mid-peer-play converges
- aboveFold useMemo no longer defeated by unstable engine object reference
- redirectWithQuery shared helper between the two /play/ redirect pages
- dropped tutorial-style WHAT comments, kept WHY comments
- JSDoc cleanup in url-utils.ts
